refactor: Separate func from Car.java
-make()
-run(List<Car> cars
-start(List<Car> cars)
